requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.4.2
google-generativeai==0.3.2
geopy==2.4.0
python-dotenv==1.0.0
gunicorn==21.2.0